Revert "test"
This reverts commit 3a6730c576e61aef992ebf7fd039823b6fb3a4d1.

+++ src/server/shared/Utilities/ServiceWin32.cpp
@@ -255,7 +255,7 @@ bool WinServiceRun()
 
     if (!StartServiceCtrlDispatcher(serviceTable))
     {
-        TC_LOG_ERROR("server.worldserver", "StartService Failed. Error [%u]", uint32(::GetLastError()));
+        TC_LOG_ERROR("server.worldserver", "StartService Failed. Error [%u]", ::GetLastError());
         return false;
     }
     return true;
